realistic_av_ready-current/requirements.txt
yara-python
python-magic-bin
pefile
oletools
pdfminer.six
tk
